ocr/requirements.txt
python-doctr[torch]
numpy==1.24.2
opencv_python==4.7.0.72
torch==1.12.1
